refactor: update package structure for authentication classes to domain.auth
