+++ glym_spring/src/main/java/glym/glym_spring/domain/auth/docs/AuthDocs.java
@@ -1,6 +1,7 @@
-package glym.glym_spring.auth.docs;
+package glym.glym_spring.domain.auth.docs;
 
 import glym.glym_spring.auth.dto.*;
+import glym.glym_spring.domain.auth.dto.*;
 import glym.glym_spring.global.dto.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import org.springframework.http.ResponseEntity;
