+++ glym_spring/src/main/java/glym/glym_spring/domain/auth/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package glym.glym_spring.auth.dto;
+package glym.glym_spring.domain.auth.dto;
 
 import lombok.Getter;
