+++ glym_spring/src/main/java/glym/glym_spring/domain/user/domain/User.java
@@ -1,6 +1,6 @@
 package glym.glym_spring.domain.user.domain;
 
-import glym.glym_spring.auth.domain.RefreshToken;
+import glym.glym_spring.domain.auth.domain.RefreshToken;
 import glym.glym_spring.domain.font.domain.FontCreation;
 import glym.glym_spring.domain.font.domain.FontProcessingJob;
 import jakarta.persistence.*;
